Database entry function
The `enter` function accepts a dictionary parameter and enters
values in the database accordingly. Handles the case for updating
an existing user as well as a new user who isn't in the system.
Can be adjusted to the parameter dictionary's format. Note, the
test code in main will do nothing unless the entries are changed,
since those specific values are now already in the database.